clamp ScreenSecurity counter at zero to prevent underflow
