+++ ios/Cove/Flows/NewWalletFlow/HotWallet/VerifyWords/VerifyWordsScreen.swift
@@ -8,26 +8,6 @@
 import SwiftUI
 import UIKit
 
-private extension View {
-    func screenshotProtected() -> some View {
-        overlay(ScreenshotProtectionView().allowsHitTesting(false))
-    }
-}
-
-private struct ScreenshotProtectionView: UIViewRepresentable {
-    func makeUIView(context _: Context) -> UIView {
-        let field = UITextField()
-        field.isSecureTextEntry = true
-        field.isUserInteractionEnabled = false
-        let view = field.layer.sublayers?.first?.delegate as? UIView ?? UIView()
-        view.translatesAutoresizingMaskIntoConstraints = false
-        view.alpha = 0
-        return view
-    }
-
-    func updateUIView(_: UIView, context _: Context) {}
-}
-
 // MARK: CONTAINER
 
 struct VerifyWordsContainer: View {
