+++ ios/Cove/Flows/SelectedWalletFlow/SecretWordsScreen.swift
@@ -8,28 +8,6 @@
 import SwiftUI
 import UIKit
 
-/// Prevents screenshots by overlaying a secure UITextField which iOS treats as protected content
-private struct ScreenshotProtection: UIViewRepresentable {
-    func makeUIView(context _: Context) -> UIView {
-        let field = UITextField()
-        field.isSecureTextEntry = true
-        field.isUserInteractionEnabled = false
-        // the secure layer UIKit attaches covers the parent view in screenshots
-        let view = field.layer.sublayers?.first?.delegate as? UIView ?? UIView()
-        view.translatesAutoresizingMaskIntoConstraints = false
-        view.alpha = 0
-        return view
-    }
-
-    func updateUIView(_: UIView, context _: Context) {}
-}
-
-private extension View {
-    func screenshotProtected() -> some View {
-        overlay(ScreenshotProtection().allowsHitTesting(false))
-    }
-}
-
 struct SecretWordsScreen: View {
     @Environment(AppManager.self) private var app
     @Environment(AuthManager.self) private var auth
